requirements.txt
numpy
tifffile
pillow
tqdm
pycine
imageio
imageio[ffmpeg]